Move metrics package one directory level up
